applied filter fix (again)

+++ bot/events/message.js
@@ -12,7 +12,7 @@ module.exports = {
 
         const loweredMsg = message.content.toLowerCase().replace(/[^a-zA-Z0-9]/g, '')
 
-        if (message.channel.type !== 'DM') {
+        if (message.guild) {
             const intuser = await handler.getUser(`${message.author.id}`)
             await intuser.modify("xp", 1, "ADD")
             const xp = await intuser.getkey("xp")
@@ -44,14 +44,18 @@ module.exports = {
         const serbianInsults = ["suka","crnja", "cigan", "crn", "jebi", "nahasasmisetarzancica", "liznesmepodmisku", "rimtutitukiđumebaokao", "sisutiskrbavu", "celavatikeva", "smrditiizustakokonjuizdupeta", "pickatiseogadilakuractiseosladio", "pucaletijebokevu", "jebemlitisirairakiju", "seremtisepodsitnozito", "posrcesmiđanicunaslamcicu", "naprskamtiseprekopudera", "jedigovnasitpapijvode", "katizveknemsamarnigovnatinecebuduslatka", "shramtebilo"]
         const serbianNsfwStuff = ["blyat","picka", "kurac", "sperme", "spermatozoid", "sperma", "vadzina", "vajijaj", "vadzajdzaj", "pičk", "dupe", "pickica"]
 
-        if (!message.member.roles.cache.has(nofilterrole)) {
-            if (englishInsults.some(s => loweredMsg.includes(s)) || englishNsfwStuff.some(s => loweredMsg.includes(s)) || serbianInsults.some(s => loweredMsg.includes(s)) || serbianNsfwStuff.some(s => loweredMsg.includes(s))) {
-                if (message.author.id === client.user.id || message.member.roles.cache.has('873680682096947210')) return;
-                message.delete()
-                message.author.send("Hey, dont be rude!").catch(() => {
-                    message.channel.send(`${message.author} dont be rude to people, people have feelings yk`)
-                })
+        if (message.member) {
+            if (!message.member.roles.cache.has(nofilterrole)) {
+                if (englishInsults.some(s => loweredMsg.includes(s)) || englishNsfwStuff.some(s => loweredMsg.includes(s)) || serbianInsults.some(s => loweredMsg.includes(s)) || serbianNsfwStuff.some(s => loweredMsg.includes(s))) {
+                    if (message.author.id === client.user.id || message.member.roles.cache.has('873680682096947210')) return;
+                    message.delete()
+                    message.author.send("Hey, dont be rude!").catch(() => {
+                        message.channel.send(`${message.author} dont be rude to people, people have feelings yk`)
+                    })
+                }
             }
+        } else {
+            //message.channel.send(`I couldn't check ${message.author}'s message due to an error, my apologies for any inconvenience caused.`)
         }
     }
 }
